Time: 0 ms (100%), Space: 8.4 MB (87.75%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int mySqrt(int x) {
-    long long low = 1;
+    int low = 1;
     long long high = x;
     
     while(low<=high){
